+++ scripts/lib/messageUserColor.js
@@ -1,4 +1,4 @@
-
+//TODO finish me
 
 /** Sourced from Dorako UI <3 */
 export async function setupMessageUserColor() {