+++ scripts/lib/notify.js
@@ -1,0 +1,1 @@
+// TODO add a module setting to give a notifcation to a suer when they referesh their Spellstrike (for tyreal wife)