+++ scripts/lib/pf2eToolbelt.js
@@ -1,61 +1,33 @@
-import { toggleCustomCSS } from "./helpers.js";
-
-export function colorizeToolbeltMessageSaves(active = true) {
-    toggleCustomCSS(colorizeToolbeltMessageSavesCSS,
-        colorizeToolbeltMessageSavesStyleID, active)
-}
-
-export function highlightToolbeltRollSaves(active = true) {
-    toggleCustomCSS(highlightToolbeltRollSavesCSS,
-        highlightToolbeltRollSavesStyleID, active)
-}
+import { CLASS, QUERY } from "./const.js";
+import { addElementsClass, removeElementsClass } from "./helpers.js";
 
 
-
-const colorizeToolbeltMessageSavesStyleID = 'colorize-toolbelt-message-saves';
-const highlightToolbeltRollSavesStyleID = 'make-toolbelt-roll-saves-glow';
-
-const colorizeToolbeltMessageSavesCSS = `
-.chat-message .message-content .pf2e-toolbelt-target-targetRows .target-row .target-header {
-
-    .controls {
-        /* reflex toolbelt*/
-        i.fa-person-running {
-            color: #F5A623;
-        }
-
-        /* will toolbelt*/
-        i.fa-brain {
-            color: #5C6BC0;
-        
-        }
-
-        /* fort toolbelt*/
-        i.fa-chess-rook {
-            color:  #43A047;
-        
-        }
-    }
-}
-`
-
-const highlightToolbeltRollSavesCSS = `
-.chat-message .message-content .pf2e-toolbelt-target-targetRows .target-row .target-header {
-	/* Changes font size of PF2e Toolbelt savings throw elements */
-	font-size: 1em;
-
-    .controls {
-
-        /* Pulsing unclicked roll */
-        i.die {
-            animation: endTurnGlow 2.5s infinite;
-        }
+export function setupColorizeToolbeltMessageSaves(active = true) {
+    const elements = document.querySelectorAll(QUERY.MESSAGE.PF2E_TOOLBELT.TARGET_ROWS);
+    if (active) {
+        addElementsClass(elements, CLASS.COLORIZE.PF2E_TOOLBELT.MESSAGE_SAVES)
+        Hooks.on('renderChatMessageHTML', colorizeToolbeltMessageSavesHook)
+    } else {
+        removeElementsClass(elements, CLASS.COLORIZE.PF2E_TOOLBELT.MESSAGE_SAVES)
+        Hooks.off('renderChatMessageHTML', colorizeToolbeltMessageSavesHook)
     }
 }
 
-@keyframes endTurnGlow {
-  0% { color: #fff; }
-  50% { color: #fd782f; }
-  100% { color: #fff; }
+export function setupHighlightToolbeltRollSaves(active = true) {
+    const elements = document.querySelectorAll(QUERY.MESSAGE.PF2E_TOOLBELT.TARGET_ROWS);
+    if (active) {
+        addElementsClass(elements, CLASS.HIGHLIGHT.PF2E_TOOLBELT.UNROLLED_ROLLS)
+        Hooks.on('renderChatMessageHTML', highlightToolbeltRollSavesHook)
+    } else {
+        removeElementsClass(elements, CLASS.HIGHLIGHT.PF2E_TOOLBELT.UNROLLED_ROLLS)
+        Hooks.off('renderChatMessageHTML', highlightToolbeltRollSavesHook)
+    }
 }
-`
+
+async function colorizeToolbeltMessageSavesHook(data, html) {
+    html.classList.add(CLASS.COLORIZE.PF2E_TOOLBELT.MESSAGE_SAVES)
+}
+
+async function highlightToolbeltRollSavesHook(data, html) {
+    html.classList.add(CLASS.HIGHLIGHT.PF2E_TOOLBELT.UNROLLED_ROLLS)
+}