+++ scripts/module.js
@@ -1,10 +1,9 @@
 import { getSetting } from "./lib/helpers.js";
 import { minifySimpleRequests } from "./lib/simpleRequests.js";
-import { setupMessageUserColor } from "./lib/messageUserColor.js";
 import { setupSettings } from "./settings.js";
-import { colorizePersistentPF2eHUD } from "./lib/pf2eHUDColorized.js";
 import { hideDefaultCraftChecks, hideSellAllTreasure } from "./lib/sheetTweaks.js";
-import { colorizeToolbeltMessageSaves, highlightToolbeltRollSaves } from "./lib/pf2eToolbelt.js";
+import { setupColorizeToolbeltMessageSaves, setupHighlightToolbeltRollSaves } from "./lib/pf2eToolbelt.js";
+import { setupColorizePersistentPF2eHUD } from "./lib/pf2eHUD.js";
 
 export const MODULE_ID = 'sundry';
 
@@ -13,21 +12,26 @@ Hooks.once('init', async function () {
 });
 
 Hooks.once('ready', async function () {
+    //Colorization
+    setupColorizeToolbeltMessageSaves(
+        getSetting("colorize.pf2e-toolbelt.target-helper.roll")
+    );
+    setupColorizePersistentPF2eHUD(
+        getSetting("colorize.pf2e-hud.persistent")
+    );
 
-    if (getSetting("colorize.pf2e-hud.persistent"))
-        colorizePersistentPF2eHUD();
+    //Highlighting
+    setupHighlightToolbeltRollSaves(
+        getSetting('highlight.pf2e-toolbelt.target-helper.roll')
+    )
 
-    if (getSetting("colorize.pf2e-toolbelt.target-helper.roll"))
-        colorizeToolbeltMessageSaves();
 
+    //Hide
     if (getSetting('hide.sell-all-treasure'))
         hideSellAllTreasure();
 
     if (getSetting('hide.default-craft-checks'))
         hideDefaultCraftChecks();
-
-    if (getSetting('highlight.pf2e-toolbelt.target-helper.roll'))
-        highlightToolbeltRollSaves();
 
     // if (getSetting('message.user-color'))
     //     setupMessageUserColor();