+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 - Add User Color to Message Header
 
+## 0.2.0
+
+- **Added**
+  - Notify a player when they perform an action that gives them a spell slot back
+- **Updated**
+  - Reworked much of the CSS for stuff to be easier to contribute to
+    - Colorize `PF2e HUD`'s `Persistent HUD`
+    - Colorize `PF2e Toolbelt`'s `Target Helper`
+    - Highlight `Pf2e Toolbelt`'s `Target Helper` unrolled rolls and make them glow
+
 ## 0.1.1
 
 - **Updated**